Feature crop (#111)
* tg_auth

* update checking

* update redis

* update redis

* update redis

* update redis

* update redis

* fix bugs

* many change

* many change

* many change

* many change

* many change

* many change

* many change

* many change

* many change

* many change

* many change

* many change

* merge

+++ todo/telbot/cleaner.py
@@ -1,5 +1,6 @@
 import re
 
+from django.conf import settings
 from telegram import Update
 from telegram.ext import CallbackContext
 
@@ -60,6 +61,9 @@ def delete_messages_by_time(chat_id: int, message_id: int) -> None:
     - message_id (:obj:`int` | :obj:`str`)
     """
     try:
-        bot.delete_message(chat_id, message_id)
+        if settings.DEBUG:
+            bot.send_message(chat_id, f'Сообщение {message_id} в чате {chat_id} будет удалено')
+        else:
+            bot.delete_message(chat_id, message_id)
     except Exception as error:
         raise KeyError(error)
